+++ lib/input/sequence_test.go
@@ -1,9 +1,11 @@
 package input
 
 import (
+	"fmt"
 	"io/ioutil"
 	"os"
 	"path/filepath"
+	"sort"
 	"testing"
 	"time"
 
@@ -75,6 +77,164 @@ consumeLoop:
 		}
 	}
 
+	assert.Equal(t, exp, act)
+
+	rdr.CloseAsync()
+	assert.NoError(t, rdr.WaitForClose(time.Second))
+}
+
+func TestSequenceJoins(t *testing.T) {
+	t.Parallel()
+
+	tmpDir, err := ioutil.TempDir("", "benthos_sequence_joins_test")
+	require.NoError(t, err)
+
+	t.Cleanup(func() { os.RemoveAll(tmpDir) })
+
+	files := map[string]string{
+		"csv1": "id,name,age\naaa,A,20\nbbb,B,21\nccc,B,22\n",
+		"csv2": "id,hobby\nccc,fencing\naaa,running\naaa,gaming\n",
+		"ndjson1": `{"id":"aaa","stuff":{"first":"foo"}}
+{"id":"bbb","stuff":{"first":"bar"}}
+{"id":"aaa","stuff":{"second":"baz"}}`,
+	}
+
+	writeFiles(t, tmpDir, files)
+
+	conf := NewConfig()
+	conf.Type = TypeSequence
+	conf.Sequence.ShardedJoin.IDField = "id"
+	conf.Sequence.ShardedJoin.Iterations = 1
+
+	csvConf := NewConfig()
+	csvConf.Type = TypeCSVFile
+	csvConf.CSVFile.Paths = []string{
+		filepath.Join(tmpDir, "csv1"),
+		filepath.Join(tmpDir, "csv2"),
+	}
+	conf.Sequence.Inputs = append(conf.Sequence.Inputs, csvConf)
+	for _, k := range []string{"ndjson1"} {
+		inConf := NewConfig()
+		inConf.Type = TypeFile
+		inConf.File.Path = filepath.Join(tmpDir, k)
+		conf.Sequence.Inputs = append(conf.Sequence.Inputs, inConf)
+	}
+
+	rdr, err := New(conf, types.NoopMgr(), log.Noop(), metrics.Noop())
+	require.NoError(t, err)
+
+	exp, act := []string{
+		`{"age":"20","hobby":["running","gaming"],"id":"aaa","name":"A","stuff":{"first":"foo","second":"baz"}}`,
+		`{"age":"21","id":"bbb","name":"B","stuff":{"first":"bar"}}`,
+		`{"age":"22","hobby":"fencing","id":"ccc","name":"B"}`,
+	}, []string{}
+
+consumeLoop:
+	for {
+		select {
+		case tran, open := <-rdr.TransactionChan():
+			if !open {
+				break consumeLoop
+			}
+			assert.Equal(t, 1, tran.Payload.Len())
+			act = append(act, string(tran.Payload.Get(0).Get()))
+			select {
+			case tran.ResponseChan <- response.NewAck():
+			case <-time.After(time.Second):
+				t.Fatalf("failed to ack after: %v", act)
+			}
+		case <-time.After(time.Second):
+			t.Fatalf("Failed to consume message after: %v", act)
+		}
+	}
+
+	sort.Strings(exp)
+	sort.Strings(act)
+	assert.Equal(t, exp, act)
+
+	rdr.CloseAsync()
+	assert.NoError(t, rdr.WaitForClose(time.Second))
+}
+
+func TestSequenceJoinsBig(t *testing.T) {
+	t.Parallel()
+
+	tmpDir, err := ioutil.TempDir("", "benthos_sequence_joins_big_test")
+	require.NoError(t, err)
+
+	t.Cleanup(func() { os.RemoveAll(tmpDir) })
+
+	jsonPath := filepath.Join(tmpDir, "one.ndjson")
+	csvPath := filepath.Join(tmpDir, "two.csv")
+
+	ndjsonFile, err := os.Create(jsonPath)
+	require.NoError(t, err)
+
+	csvFile, err := os.Create(csvPath)
+	require.NoError(t, err)
+
+	conf := NewConfig()
+	conf.Type = TypeSequence
+	conf.Sequence.ShardedJoin.IDField = "id"
+	conf.Sequence.ShardedJoin.Iterations = 5
+
+	csvConf := NewConfig()
+	csvConf.Type = TypeCSVFile
+	csvConf.CSVFile.Paths = []string{csvPath}
+	conf.Sequence.Inputs = append(conf.Sequence.Inputs, csvConf)
+
+	jsonConf := NewConfig()
+	jsonConf.Type = TypeFile
+	jsonConf.File.Paths = []string{jsonPath}
+	jsonConf.File.Codec = "lines"
+	conf.Sequence.Inputs = append(conf.Sequence.Inputs, jsonConf)
+
+	totalRows := 1000
+
+	exp, act := []string{}, []string{}
+
+	_, err = csvFile.Write([]byte("id,bar\n"))
+	require.NoError(t, err)
+	for i := 0; i < totalRows; i++ {
+		exp = append(exp, fmt.Sprintf(`{"bar":["bar%v","baz%v"],"foo":"foo%v","id":"%v"}`, i, i, i, i))
+
+		_, err = fmt.Fprintf(ndjsonFile, "{\"id\":\"%v\",\"foo\":\"foo%v\"}\n", i, i)
+		require.NoError(t, err)
+
+		_, err = fmt.Fprintf(csvFile, "%v,bar%v\n", i, i)
+		require.NoError(t, err)
+	}
+	for i := 0; i < totalRows; i++ {
+		_, err = fmt.Fprintf(csvFile, "%v,baz%v\n", i, i)
+		require.NoError(t, err)
+	}
+	require.NoError(t, ndjsonFile.Close())
+	require.NoError(t, csvFile.Close())
+
+	rdr, err := New(conf, types.NoopMgr(), log.Noop(), metrics.Noop())
+	require.NoError(t, err)
+
+consumeLoop:
+	for {
+		select {
+		case tran, open := <-rdr.TransactionChan():
+			if !open {
+				break consumeLoop
+			}
+			assert.Equal(t, 1, tran.Payload.Len())
+			act = append(act, string(tran.Payload.Get(0).Get()))
+			select {
+			case tran.ResponseChan <- response.NewAck():
+			case <-time.After(time.Second):
+				t.Fatalf("failed to ack after: %v", act)
+			}
+		case <-time.After(time.Second):
+			t.Fatalf("Failed to consume message after: %v", act)
+		}
+	}
+
+	sort.Strings(exp)
+	sort.Strings(act)
 	assert.Equal(t, exp, act)
 
 	rdr.CloseAsync()